+++ resources/release-notes.md
@@ -1,4 +1,5 @@
 # App
 
+- Add placeholders for CPT settings with dynamic defaults
 - Disable "Save Figures" and "Close Figures" buttons after using "Close Figures"
 - Fix issues with validation of CPT primary and secondary modes, and range patterns (`mag.app.control.CPT`)